backend/requirements.txt
# Core Dependencies
openai>=1.12.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0
aiohttp>=3.9.0

# Terminal UI
rich>=13.7.0
textual>=0.47.0
prompt-toolkit>=3.0.43

# Web Framework
fastapi>=0.110.0
python-multipart>=0.0.9
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Browser Automation
playwright>=1.40.0
selenium>=4.15.0

# Telegram
telethon>=1.33.0
python-telegram-bot>=20.7

# Social Media APIs
tweepy>=4.14.0
linkedin-api>=2.2.0

# Text Processing
beautifulsoup4>=4.12.0
lxml>=4.9.0
html5lib>=1.1

# Contact Extraction
phonenumbers>=8.13.0
email-validator>=2.1.0

# Natural Language Processing
spacy>=3.7.0
nltk>=3.8.1

# Async Support
asyncio>=3.4.3
aiofiles>=23.2.1

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
colorama>=0.4.6
tqdm>=4.66.0

# Logging
loguru>=0.7.2

# Data Processing
pandas>=2.1.0
numpy>=1.26.0

# Scheduling
apscheduler>=3.10.0

# HTTP Client
httpx>=0.25.0

# JSON Processing
jsonschema>=4.20.0

# Encryption

# Rate Limiting
ratelimit>=2.2.1

# Environment
environs>=10.0.0
# ── FastAPI + Server ───────────────────────────────────────────────────────────
uvicorn[standard]>=0.27.0
pyyaml>=6.0.1
httpx>=0.27.0

# ── Database ───────────────────────────────────────────────────────────────────
asyncpg>=0.29.0           # Async PostgreSQL (platform DB)
psycopg2-binary>=2.9.9    # Sync PostgreSQL (migrations/tools)
# SQLite is built-in — used for per-user agent data

# ── Auth ───────────────────────────────────────────────────────────────────────
google-auth>=2.28.0
PyJWT>=2.8.0
cryptography>=42.0.0

# ── Media processing ───────────────────────────────────────────────────────────
pillow>=10.2.0
opencv-python-headless>=4.9.0    # Video frame extraction (headless = no GUI deps)
prompt_toolkit>=3.0.43

# ── Utilities ─────────────────────────────────────────────────────────────────
pydantic>=2.6.0
qrcode>=7.4.2